Move AttributeData into separate file
