+++ src/InputHandler.test.ts
@@ -8,8 +8,10 @@ import { InputHandler } from './InputHandler';
 import { MockInputHandlingTerminal, TestTerminal } from './TestUtils.test';
 import { Terminal } from './Terminal';
 import { IBufferLine } from 'common/Types';
-import { Attributes, AttributeData, DEFAULT_ATTR_DATA } from 'common/buffer/BufferLine';
+import { DEFAULT_ATTR_DATA } from 'common/buffer/BufferLine';
 import { CellData } from 'common/buffer/CellData';
+import { Attributes } from 'common/buffer/Constants';
+import { AttributeData } from 'common/buffer/AttributeData';
 
 describe('InputHandler', () => {
   describe('save and restore cursor', () => {
